requirements.txt
cerebras_cloud_sdk
aider-install
